+++ src/App.tsx
@@ -1,4 +1,4 @@
-import React, {useEffect, useState, createContext }  from "react"
+import React, {useEffect, useState, useContext }  from "react"
 import { BrowserRouter, Routes, Route, Navigate }  from "react-router-dom"
 
 import { CommonLayout } from "components/layouts/CommonLayout"
@@ -11,63 +11,14 @@ import { Ship } from "components/pages/Ship"
 
 import { getCurrentUser } from "lib/api/auth"
 import { User } from "interfaces/index"
+import { AuthContextProvider } from "components/providers/AuthContextProvider"
 
-
-export const AuthContext = createContext({} as {
-  loading: boolean
-  setLoading: React.Dispatch<React.SetStateAction<boolean>>
-  isSignedIn: boolean
-  setIsSignedIn: React.Dispatch<React.SetStateAction<boolean>>
-  currentUser: User | undefined
-  setCurrentUser: React.Dispatch<React.SetStateAction<User | undefined>>
-})
 
 function App() {
-  const [loading, setLoading] = useState<boolean>(true)
-  const [isSignedIn, setIsSignedIn] = useState<boolean>(false)
-  const [currentUser, setCurrentUser] = useState<User | undefined>()
-
-  const handleGetCurrentUser = async () => {
-    try {
-      const res = await getCurrentUser()
-      console.log(res)
-
-      if (res?.status === 200) {
-        setIsSignedIn(true)
-        setCurrentUser(res?.data.currentUser)
-      }
-      else {
-        console.log("No current User")
-      }
-    }
-    catch(err) {
-      console.log(err)
-    }
-
-    setLoading(false)
-  }
-
-  useEffect(() => {
-    handleGetCurrentUser()
-  }, [setCurrentUser])
-
-  const Private =({ children }: {children: React.ReactElement }) => {
-    if(!loading) {
-      if(isSignedIn) {
-        return children
-      }
-      else {
-        return <Navigate to ="/signin"/>
-      }
-    }
-    else {
-      return<></>
-    }
-  }
 
   return (
     <BrowserRouter>
-      <AuthContext.Provider value={{ loading, setLoading, isSignedIn, setIsSignedIn, currentUser, setCurrentUser }}>
+      <AuthContextProvider>
         <CommonLayout>
           <Routes>
             <Route  path="/signup" element={<SignUp />} />
@@ -78,7 +29,7 @@ function App() {
             <Route  path="/ship" element={<Ship />} />
           </Routes>
         </CommonLayout>
-      </AuthContext.Provider>
+      </AuthContextProvider>
     </BrowserRouter>
   );
 }
