+++ src/components/pages/Home.tsx
@@ -1,6 +1,6 @@
 import React, { useContext } from "react"
 
-import { AuthContext } from "App"
+import { AuthContext } from "components/providers/AuthContextProvider"
 
 export const Home: React.FC = () =>  {
   const { isSignedIn, currentUser } = useContext(AuthContext)
