+++ src/components/providers/AuthContextProvider.tsx
@@ -1,0 +1,59 @@
+import { createContext, useState, useEffect } from "react";
+import { User } from "interfaces/index"
+import { getCurrentUser } from "lib/api/auth"
+
+
+
+export const AuthContext = createContext({} as {
+  loading: boolean
+  setLoading: React.Dispatch<React.SetStateAction<boolean>>
+  isSignedIn: boolean
+  setIsSignedIn: React.Dispatch<React.SetStateAction<boolean>>
+  currentUser: User | undefined
+  setCurrentUser: React.Dispatch<React.SetStateAction<User | undefined>>
+});
+
+
+
+interface props {
+  children: React.ReactNode;
+}
+
+
+export const AuthContextProvider = (props: props) => {
+
+  const [loading, setLoading] = useState<boolean>(true)
+  const [isSignedIn, setIsSignedIn] = useState<boolean>(false)
+  const [currentUser, setCurrentUser] = useState<User | undefined>()
+
+  const handleGetCurrentUser = async () => {
+    try {
+
+      const res = await getCurrentUser()
+      console.log(res)
+
+      if (res?.status === 200) {
+        setIsSignedIn(true)
+        setCurrentUser(res?.data.currentUser)
+      }
+      else {
+        console.log("No current User")
+      }
+    }
+    catch(err) {
+      console.log(err)
+    }
+
+    setLoading(false)
+  }
+
+  useEffect(() => {
+    handleGetCurrentUser()
+  }, [setCurrentUser])
+
+  return (
+    <AuthContext.Provider value={{ loading, setLoading, isSignedIn, setIsSignedIn, currentUser, setCurrentUser }}>
+      {props.children}
+    </AuthContext.Provider>
+  )
+}
